// src/api/rathena/DatabaseDriver.types.ts
export type LoginEntity = zod.infer<typeof LoginEntityType>;
export const LoginEntityType = zod.object({
  account_id: zod.number(),
  birthdate: zod.date().nullable(),
  character_slots: zod.number(),
  email: zod.string(),
  expiration_time: zod.number(),
  group_id: zod.number(),
  last_ip: zod.string(),
  lastlogin: zod.date().nullable(),
  logincount: zod.number(),
  old_group: zod.number(),
  pincode: zod.string(),
  pincode_change: zod.number(),
  sex: zod.number(),
  state: zod.number(),
  unban_time: zod.number(),
  user_pass: zod.string(),
  userid: zod.string(),
  vip_time: zod.number(),
  web_auth_token: zod.string().nullable(),
  web_auth_token_enabled: zod.number(),
});